refactor: extract shared find_device_entity helper (#2101)
* refactor: extract shared find_device_entity helper

The identical nested find_entity() closure existed in switch.py (child-lock
lookup) and model_fixes/default.py (initial tweak lookups). Extracted the
matching logic into find_device_entity() in utils/helpers.py and routed both
call sites through it. No behaviour change.

Deliberately left the similar _get_active_algorithms() in sensor.py and
config_flow.py separate: they consume different inputs (Trv objects vs raw
config dicts), live in different layers, and only the sensor variant logs a
per-TRV warning; unifying them would couple the layers behind a callback for
marginal benefit.

* refactor: harden find_device_entity against one-shot iterables

Materialize domains/keywords to tuples once and lower-case keywords so the
helper works with generator inputs and case-insensitive keyword matching, not
just the lowercase list literals the current call sites happen to pass.

* chore: drop issue references from convert_to_float comment
